Fix a bug when uncommitted records may become visible after the database sweep
When the server crash happens, the database might have pointer pages with flags which are not consistent with flags on data pages. If ppg_dp_swept is set for a data page which does not actually have dpg_swept set, then records on this page won't be garbage-collected, and they will become visible after OIT is advanced at the end of the sweep. The same problem with ppg_dp_empty because the sweep skips such pages as well.

Key changes:
1. If ppg_dp_swept or ppg_dp_empty is cleared for DPM_primary, a pointer page is always written before a data page.
2. When an extent is allocated, ppg_dp_empty is not set for its data pages. This is needed to avoid performance penalty due to precedence circles.
3. Data page numbers of a newly allocated extent are temporary saved in RelationPages to keep bulk inserts functional.

Thanks to Vlad Khorsun (@hvlad) for the initial idea of the fix.